requirements.txt
pandas~=1.2.1
requests~=2.25.1
openpyxl~=3.0.6
GoogleAppEngineCloudStorageClient~=1.9.22.1
google-cloud-storage~=1.35.1